Have fully typed contract write arguments

+++ apps/web/components/IncreasePower.tsx
@@ -155,7 +155,7 @@ export const IncreasePower = ({
   } = useContractWriteWithConfirmations({
     ...registryContractCallConfig,
     functionName: "increasePower",
-    args: [requestedAmount as bigint],
+    args: [requestedAmount],
     showNotification: false,
     onConfirmations: () => {
       publish({
@@ -177,7 +177,7 @@ export const IncreasePower = ({
   } = useContractWriteWithConfirmations({
     ...registryContractCallConfig,
     functionName: "decreasePower",
-    args: [requestedAmount as bigint],
+    args: [requestedAmount],
     fallbackErrorMessage: "Error decreasing power. Please try again.",
     onConfirmations: () => {
       publish({
